backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic-settings==2.7.1
supabase==2.11.0
openai==1.58.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
